requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
boto3==1.28.73
botocore==1.31.73
distlib==0.3.7
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
filelock==3.13.0
jmespath==1.0.1
mysqlclient==2.2.0
netifaces==0.11.0
platformdirs==3.11.0
python-dateutil==2.8.2
pytz==2023.3.post1
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
virtualenv==20.24.6
Jinja2 
pdfkit